KER-210 fix merge conflicts
